switched to java 1.6

+++ cee-news-webreader/src/main/java/com/cee/news/client/workingset/WorkingSetService.java
@@ -21,7 +21,7 @@ public interface WorkingSetService extends RemoteService {
 		}
 	}
 	
-	int getWorkingSetCount();
+	long getWorkingSetCount();
 	
 	List<String> getWorkingSetsOrderedByName();
